rename CmsPageCategory to BreadcrumbTreeItem and move to QueryResult
