+++ src/Adapter/CMS/PageCategory/QueryHandler/GetCmsPageCategoriesForBreadcrumbHandler.php
@@ -33,7 +33,7 @@ use PrestaShop\PrestaShop\Core\Domain\CmsPageCategory\Exception\CmsPageCategoryE
 use PrestaShop\PrestaShop\Core\Domain\CmsPageCategory\Exception\CmsPageCategoryNotFoundException;
 use PrestaShop\PrestaShop\Core\Domain\CmsPageCategory\Query\GetCmsPageCategoriesForBreadcrumb;
 use PrestaShop\PrestaShop\Core\Domain\CmsPageCategory\QueryHandler\GetCmsPageCategoriesForBreadcrumbHandlerInterface;
-use PrestaShop\PrestaShop\Core\Domain\CmsPageCategory\ValueObject\CmsPageCategory;
+use PrestaShop\PrestaShop\Core\Domain\CmsPageCategory\QueryResult\BreadcrumbTreeItem;
 use PrestaShopException;
 
 /**
@@ -100,7 +100,7 @@ final class GetCmsPageCategoriesForBreadcrumbHandler implements GetCmsPageCatego
 
         if (CmsPageRootCategorySettings::ROOT_CMS_PAGE_CATEGORY_ID === $query->getCurrentCategoryId()->getValue()) {
             return new BreadcrumbTree([
-                new CmsPageCategory(
+                new BreadcrumbTreeItem(
                     (int) $rootCategoryData['id_cms_category'],
                     $rootCategoryData['name']
                 ),
@@ -113,7 +113,7 @@ final class GetCmsPageCategoriesForBreadcrumbHandler implements GetCmsPageCatego
 
         $categories = [];
         foreach ($parentCategories as $category) {
-            $categories[] = new CmsPageCategory(
+            $categories[] = new BreadcrumbTreeItem(
                 (int) $category['id_cms_category'],
                 $category['name']
             );
